+++ source/src/enet_encap/endianconv.h
@@ -57,8 +57,8 @@ CipUdint GetUdintFromMessage(const CipOctet **const buffer_address);
  * @param data value to be written
  * @param buffer pointer where data should be written.
  */
-int AddSintToMessage(const EipUint8 data,
-                     EipUint8 **const buffer);
+size_t AddSintToMessage(const EipUint8 data,
+                        EipUint8 **const buffer);
 
 /** @ingroup ENCAP
  *
@@ -68,8 +68,8 @@ int AddSintToMessage(const EipUint8 data,
  *
  * @return Length in bytes of the encoded message
  */
-int AddIntToMessage(const EipUint16 data,
-                    EipUint8 **const buffer);
+size_t AddIntToMessage(const EipUint16 data,
+                       EipUint8 **const buffer);
 
 /** @ingroup ENCAP
  *
@@ -79,8 +79,8 @@ int AddIntToMessage(const EipUint16 data,
  *
  * @return Length in bytes of the encoded message
  */
-int AddDintToMessage(const EipUint32 data,
-                     EipUint8 **const buffer);
+size_t AddDintToMessage(const EipUint32 data,
+                        EipUint8 **const buffer);
 
 #ifdef OPENER_SUPPORT_64BIT_DATATYPES
 
@@ -94,8 +94,8 @@ EipUint64 GetLintFromMessage(const EipUint8 **const buffer);
  *
  * @return Length in bytes of the encoded message
  */
-int AddLintToMessage(const EipUint64 pa_unData,
-                     EipUint8 **const buffer);
+size_t AddLintToMessage(const EipUint64 pa_unData,
+                        EipUint8 **const buffer);
 
 #endif
